Refactor cli input to own module
