+++ src/lib.rs
@@ -2,6 +2,6 @@ pub mod avl_tree;
 pub mod rb_tree;
 pub mod bs_tree;
 
-pub use avl_tree::AVL;
-pub use rb_tree::RedBlack;
+pub use avl_tree::AVLTree;
+pub use rb_tree::RedBlackTree;
 pub use bs_tree::BinarySearchTree;
